+++ database/migrations/2021_01_09_173316_create_wyplata_table.php
@@ -17,7 +17,7 @@ class CreateWyplataTable extends Migration
             $table->id('numer dowodu wplaty');
             $table->date('data');
             $table->text('tresc');
-            $table->integer('kwota rozchodu');
+            $table->decimal('kwota rozchodu', 10, 2);
         });
     }
